Undo Deletion
This step should have been done on GitKraken however I attempted to figure it out for two hours and could not do it (I could not simply click the undo button to undo step three). Because of this, I used GitHub's history section, copied the deleted section, and pasted it into my file.