Modification 6 - Add Copyright Information
Determination of copyright based upon source consulted for additional paragraphs/title information is displayed below.

From Project Gutenberg:
This eBook is for the use of anyone anywhere in the United States and most
other parts of the world at no cost and with almost no restrictions
whatsoever.  You may copy it, give it away or re-use it under the terms of
the Project Gutenberg License included with this eBook or online at
www.gutenberg.org.  If you are not located in the United States, you'll have
to check the laws of the country where you are located before using this ebook.

In Canada, The Iliad of Homer is in the public domain.